+++ library-build-transformer/src/main/kotlin/io/realm/buildtransformer/util/Stopwatch.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package buildtransformer.util
+package io.realm.buildtransformer.util
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
